refactor: move tag regexp to constants.ts
