+++ apps/editor/src/convertors/toWysiwyg/htmlToWwConvertors.ts
@@ -1,31 +1,13 @@
 import { MdNode } from '@toast-ui/toastmark';
 import { sanitizeXSSAttributeValue } from '@/sanitizer/htmlSanitizer';
-import { includes } from '@/utils/common';
 
 import {
   HTMLToWwConvertorMap,
   FlattenHTMLToWwConvertorMap,
   ToWwConvertorState,
 } from '@t/convertor';
-
-const TAG_NAME = '[A-Za-z][A-Za-z0-9-]*';
-const ATTRIBUTE_NAME = '[a-zA-Z_:][a-zA-Z0-9:._-]*';
-const UNQUOTED_VALUE = '[^"\'=<>`\\x00-\\x20]+';
-
-const SINGLE_QUOTED_VALUE = "'[^']*'";
-const DOUBLE_QUOTED_VALUE = '"[^"]*"';
-
-const ATTRIBUTE_VALUE = `(?:${UNQUOTED_VALUE}|${SINGLE_QUOTED_VALUE}|${DOUBLE_QUOTED_VALUE})`;
-const ATTRIBUTE_VALUE_SPEC = `${'(?:\\s*=\\s*'}${ATTRIBUTE_VALUE})`;
-
-export const ATTRIBUTE = `${'(?:\\s+'}${ATTRIBUTE_NAME}${ATTRIBUTE_VALUE_SPEC}?)`;
-
-const OPEN_TAG = `<(${TAG_NAME})(${ATTRIBUTE})*\\s*/?>`;
-const CLOSE_TAG = `</(${TAG_NAME})\\s*[>]`;
-
-const HTML_TAG = `(?:${OPEN_TAG}|${CLOSE_TAG})`;
-
-export const reHTMLTag = new RegExp(`^${HTML_TAG}`, 'i');
+import { includes } from '@/utils/common';
+import { reHTMLTag } from '@/utils/constants';
 
 export function getTextWithoutTrailingNewline(text: string) {
   return text[text.length - 1] === '\n' ? text.slice(0, text.length - 1) : text;
